Revert "Abstract basic auth in AppEngine APIs out of receiver package (#634)"
This reverts commit 2453ca7a16c358b2b6a790e2e15e8a73a579bc5c.
