+++ api/receiver/appengine.go
@@ -12,7 +12,6 @@ import (
 	"net/http"
 	"net/url"
 
-	"github.com/web-platform-tests/wpt.fyi/api/auth"
 	"github.com/web-platform-tests/wpt.fyi/shared"
 	"google.golang.org/appengine/datastore"
 	"google.golang.org/appengine/taskqueue"
@@ -30,9 +29,9 @@ type DatastoreKey struct {
 
 // AppEngineAPI abstracts all AppEngine APIs used by the results receiver.
 type AppEngineAPI interface {
-	auth.AppEngineAPI
-
+	Context() context.Context
 	AddTestRun(testRun *shared.TestRun) (*DatastoreKey, error)
+	AuthenticateUploader(username, password string) bool
 	// The three methods below are exported for webapp.admin_handler.
 	IsLoggedIn() bool
 	IsAdmin() bool
@@ -47,8 +46,6 @@ type AppEngineAPI interface {
 
 // appEngineAPIImpl is backed by real AppEngine APIs.
 type appEngineAPIImpl struct {
-	auth.AppEngineAPI
-
 	ctx    context.Context
 	client *http.Client
 	gcs    gcs
@@ -58,20 +55,13 @@ type appEngineAPIImpl struct {
 // NewAppEngineAPI creates a real AppEngineAPI from a given context.
 func NewAppEngineAPI(ctx context.Context) AppEngineAPI {
 	return &appEngineAPIImpl{
-		AppEngineAPI: auth.NewAppEngineAPI(ctx),
-		ctx:          ctx,
-		queue:        ResultsQueue,
+		ctx:   ctx,
+		queue: ResultsQueue,
 	}
 }
 
-// NewAppEngineAPIWithAuth creates a real AppEngineAPI from a given context and
-// authentication API.
-func NewAppEngineAPIWithAuth(ctx context.Context, aeAuth auth.AppEngineAPI) AppEngineAPI {
-	return &appEngineAPIImpl{
-		AppEngineAPI: aeAuth,
-		ctx:          ctx,
-		queue:        ResultsQueue,
-	}
+func (a *appEngineAPIImpl) Context() context.Context {
+	return a.ctx
 }
 
 func (a *appEngineAPIImpl) AddTestRun(testRun *shared.TestRun) (*DatastoreKey, error) {
@@ -84,6 +74,15 @@ func (a *appEngineAPIImpl) AddTestRun(testRun *shared.TestRun) (*DatastoreKey, e
 		Kind: key.Kind(),
 		ID:   key.IntID(),
 	}, nil
+}
+
+func (a *appEngineAPIImpl) AuthenticateUploader(username, password string) bool {
+	key := datastore.NewKey(a.ctx, "Uploader", username, 0, nil)
+	var uploader shared.Uploader
+	if err := datastore.Get(a.ctx, key, &uploader); err != nil || uploader.Password != password {
+		return false
+	}
+	return true
 }
 
 func (a *appEngineAPIImpl) IsLoggedIn() bool {
